apps/nlp/model/pyproject.toml
[tool.coverage.run]
branch = true
source = ["model"]

[tool.coverage.report]
exclude_lines = ["if TYPE_CHECKING:"]
show_missing = true

[tool.pytest.ini_options]
addopts = "--cov --cov-report html:'../../coverage//apps/model/html' --cov-report xml:'../../coverage//apps/model/coverage.xml' --html='../../reports//apps/model/unittests/html/index.html' --junitxml='../../reports//apps/model/unittests/junit.xml'"

[tool.poetry]
name = "model"
version = "1.0.0"
description = "Train and evaluate a model"
authors = ["Hutech University <hutech@hutech.edu.vn>"]
license = "MIT"
group = {}

[[tool.poetry.packages]]
include = "model"

[tool.poetry.dependencies]
python = ">=3.9 <=3.12"
torch = "^2.1.1"
transformers = "^4.35.2"
datasets = "^2.15.0"
python-dotenv = "^1.0.0"
pydantic = "^2.5.2"

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
line_length = 88

[tool.nx]
autoActivate = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
